@@ -2,16 +2,15 @@
 
 package world.ofunny.bpmproxy.Events;
 
-import net.md_5.bungee.api.event.PostLoginEvent;
-import net.md_5.bungee.api.event.ServerConnectEvent;
-import net.md_5.bungee.api.plugin.Listener;
-import net.md_5.bungee.event.EventHandler;
-import net.md_5.bungee.event.EventPriority;
+import com.velocitypowered.api.event.Subscribe;
+import com.velocitypowered.api.event.connection.PostLoginEvent;
+import com.velocitypowered.api.event.player.ServerConnectedEvent;
+import com.velocitypowered.api.event.player.ServerPreConnectEvent;
 import world.ofunny.bpmproxy.Module.ForcedHostsModule;
 import world.ofunny.bpmproxy.Module.LuckPermsModule;
 import world.ofunny.bpmproxy.config.Config;
 
-public class EventListener implements Listener {
+public class EventListener {
 
 	/*
 	 * Initialisation
@@ -37,7 +36,7 @@ public class EventListener implements Listener {
 	/**
 	 * Event called as soon as a connection has a ProxiedPlayer and is ready to be connected to a server.
 	 */
-    @EventHandler(priority = EventPriority.LOWEST)
+    @Subscribe(priority = 10000)
     public void onPostLoginEvent(PostLoginEvent event) {
 
     	/*
@@ -51,8 +50,8 @@ public class EventListener implements Listener {
 	 * Called when deciding to connect to a server. At the time when this event is called, no connection has actually been made.
 	 * Cancelling the event will ensure that the connection does not proceed and can be useful to prevent certain players from accessing certain servers.
 	 */
-	@EventHandler(priority = EventPriority.LOWEST)
-	public void onServerConnectEvent(ServerConnectEvent serverConnectEvent) {
+	@Subscribe(priority = 10000)
+	public void onServerConnectEvent(ServerPreConnectEvent serverConnectEvent) {
 
 		/*
 		 * Perform the group change if the permission module has been activated.
