@@ -1,7 +1,9 @@
 package world.ofunny.bpmproxy.Module;
 
-import net.md_5.bungee.api.ProxyServer;
-import net.md_5.bungee.api.plugin.PluginManager;
+import com.velocitypowered.api.command.CommandManager;
+import com.velocitypowered.api.command.CommandMeta;
+import com.velocitypowered.api.plugin.PluginManager;
+import world.ofunny.bpmproxy.BedrockPlayerManagerProxy;
 import world.ofunny.bpmproxy.Utils.Logger;
 import world.ofunny.bpmproxy.commands.PlayerTransferCommand;
 import world.ofunny.bpmproxy.config.Config;
@@ -34,7 +36,7 @@ public enum CommandModule {
      * Registers all commands used within this plugin.
      */
     public void registerAllCommands() {
-        PluginManager pluginManager = ProxyServer.getInstance().getPluginManager();
+        CommandManager commandManager = BedrockPlayerManagerProxy.getInstance().getProxyServer().getCommandManager();
         config.getPlayerTransferCommands().forEach((serverName, commandAliasList) -> {
             // Iterate trough every alias name in the list of the given server key.
             commandAliasList.forEach((String commandAlias) -> {
@@ -42,7 +44,11 @@ public enum CommandModule {
                 if(commandAlias.isEmpty()) {
                     logger.logWarning("Registration of player transfer alias command for server "+serverName+" failed – a command name is blank or invalid, check your config.yml!");
                 } else {
-                    pluginManager.registerCommand(config.getPlugin(), new PlayerTransferCommand(serverName, commandAlias, ""));
+                    CommandMeta meta = commandManager.metaBuilder(commandAlias)
+                            .plugin(BedrockPlayerManagerProxy.getInstance())
+                            .build();
+
+                    commandManager.register(meta, new PlayerTransferCommand(serverName, ""));
                     logger.debugLogInfo("Registered player transfer alias command "+commandAlias+" for server "+serverName+"!");
                 }// end if command not blank
             });// end for each alias in the current list of aliases
